+++ rbx_dom_weak/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added conversions from `BrickColor` to `Color3` and `Color3uint8` ([#46](https://github.com/LPGhatguy/rbx-dom/pull/46))
+* Changed the human readable serialization of `BinaryString` values to be base64-encoded. This makes JSON-encoded values much smaller. This changes the details of the unspecified serialization of rbx_dom_weak.
 
 ## 1.6.0 (2019-05-12)
 * Added `RbxTree::move_instance` API to move instances from one tree to another.